% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_symbolic.R
\name{as-symbolic}
\alias{as-symbolic}
\alias{as_symbolic_obj}
\alias{symblc}
\alias{as_symbolic_objs}
\alias{symblcs}
\title{Object(s) as symbolics}
\usage{
as_symbolic_obj(x, with_ns = TRUE)

symblc(x, with_ns = TRUE)

as_symbolic_objs(
  ...,
  .x = NULL,
  .named = FALSE,
  .name_all = FALSE,
  .names_type = c("full", "sym", "head")
)

symblcs(
  ...,
  .x = NULL,
  .named = FALSE,
  .name_all = FALSE,
  .names_type = c("full", "sym", "head")
)
}
\arguments{
\item{x}{an R object. Precisely, a syntactic literal, a symbol, a call, a function, or a formula.}

\item{...}{R objects as above.}

\item{.x}{a list of R objects. It gets appended to \code{...}.}

\item{.named}{TRUE or FALSE. If \code{TRUE} (default), the unnamed elements of the output of \code{as_symbolic_objs} are automatically named.}

\item{.name_all}{TRUE or FALSE. If \code{TRUE} (default), all elements of the output of \code{as_symbolic_objs} are automatically named.}

\item{.names_type}{one of \code{c("full", "sym", "head")}.
\describe{
\item{"full"}{The output is equivalent to \code{deparse(x)}.}
\item{"sym"}{The function returns the string equivalent of a symbol \code{x} is a symbol or a call with no arguments.}
\item{"head"}{The function returns only the string equivalent of a symbol or the call name.}
}}
}
\value{
a symbolic object. That is, a symbol or a call.
}
\description{
These functions take various R objects and return the equivalent symbols orcalls (e.g. \code{"func"} to \code{func}, \code{"pkg::func(x, ...)"} to \code{pkg::func(x, ...)}, etc.`).
\itemize{
\item \code{\link[=as_symbolic_obj]{as_symbolic_obj()}} returns a symbolic.
\item \code{\link[=as_symbolic_objs]{as_symbolic_objs()}} returns a list of symbolics if length of input greater than one. A symbolic otherwise.
}
}
\examples{
as_symbolic_obj(quote(mean))
as_symbolic_obj("mean")
as_symbolic_obj("base::mean(1:10)")
as_symbolic_obj(quote(sd(1:10)))
as_symbolic_obj(quote(base::mean))

as_symbolic_objs(quote(mean), "mean", "base::mean(1:10)", quote(sd(1:10)))

x <- list(quote(mean), b = "mean", c = "base::mean(1:10)", quote(sd(1:10)))
as_symbolic_objs("fn", .x = x)
as_symbolic_objs("fn", .x = x, .named = TRUE)
as_symbolic_objs("fn", .x = x, .named = TRUE, .name_all = TRUE)
as_symbolic_objs("fn", .x = x, .named = TRUE, .name_all = TRUE, .names_type = "full")
as_symbolic_objs("fn", .x = x, .named = TRUE, .name_all = TRUE, .names_type = "sym")
as_symbolic_objs("fn", .x = x, .named = TRUE, .name_all = TRUE, .names_type = "head")
}
