% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_missing.R
\name{install_missing}
\alias{install_missing}
\title{Write environment variables to the project .Renviron-file}
\usage{
install_missing(.pkg_cran = character(0L), .pkg_gith = character(0L))
}
\arguments{
\item{.pkg_cran}{...}

\item{.pkg_gith}{...}
}
\value{
invisible \code{NULL}
}
\description{
\code{install_missing} - Install missing packages from CRAN and/or GitHub
}
\examples{
install_missing
}
