% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_test.R
\name{is_numeric_chr}
\alias{is_numeric_chr}
\title{Does a character vector only contains numbers?}
\usage{
is_numeric_chr(x, na.omit = TRUE)
}
\arguments{
\item{x}{a character vector or an object coercible to one.}

\item{na.omit}{TRUE or FALSE. If TRUE (default), NAs are omitted.}
}
\value{
TRUE or FALSE.
}
\description{
Does a character vector only contains numbers?
}
