% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types_test.R
\name{namespaced-language}
\alias{namespaced-language}
\alias{is_ns_sym}
\alias{is_ns_call}
\title{Does a language object include namespace?}
\usage{
is_ns_sym(x, ns = NULL, private = NULL)

is_ns_call(x, ns = NULL, private = NULL)
}
\arguments{
\item{x}{an R object.}

\item{ns}{a string indicating a namespace name.}

\item{private}{NULL, TRUE, or FALSE.
\describe{
\item{NULL (default)}{The function returns TRUE whether the object is exported from the namespace or not.}
\item{TRUE}{The function returns TRUE if it represents an object not exported from the namespace (i.e. \code{:::}).}
\item{FALSE}{The function returns TRUE if it represents an object not exported from the namespace (i.e. \code{::}).}
}}
}
\value{
TRUE or FALSE.
}
\description{
Does a language object include namespace?
}
