% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep.R
\name{rep-helpers}
\alias{rep-helpers}
\alias{rep_times}
\alias{rep_each}
\title{Helpers for \link[base:rep]{base::rep}}
\usage{
rep_times(x, times, len = NULL)

rep_each(x, times, len = NULL)
}
\arguments{
\item{x}{a vector (of any mode including a list) or a factor or (for rep only) a POSIXct or POSIXlt or Date object; or an S4 object containing such an object.}

\item{times}{an integer-valued vector giving the (non-negative) number of times to repeat each element if of length length(x), or to repeat the whole vector if of length 1. Negative or NA values are an error. A double vector is accepted, other inputs being coerced to an integer or double vector.}

\item{len}{non-negative numeric coercible to integer.}
}
\value{
a vector.
}
\description{
Helpers for \link[base:rep]{base::rep}
}
