% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setenv_persist.R
\name{setenv_persist}
\alias{setenv_persist}
\title{Write environment variables to the project .Renviron-file}
\usage{
setenv_persist(..., .path_proj = here::here())
}
\arguments{
\item{...}{...}

\item{.path_proj}{...}
}
\value{
invisible \code{NULL}
}
\description{
\code{setenv_persist} - Convert all Emojis to some ...
}
\examples{
setenv_persist
}
