% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_enum}
\alias{str_enum}
\title{Enumerate and format the elements of a character vector.}
\usage{
str_enum(x, sep = ": ", pad_num = TRUE, format_str = TRUE)
}
\arguments{
\item{x}{a character vector.}

\item{sep}{a string indicating the separation betweem item numbers and strings.}

\item{pad_num}{TRUE or FALSE. If TRUE (default), numbers are padded with str_pad_num.}

\item{format_str}{TRUE or FALSE. If TRUE (default), strings are formatted with str_format.}
}
\value{
a character vector.
}
\description{
Enumerate and format the elements of a character vector.
}
