% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_format}
\alias{str_format}
\title{String formatiing/padding}
\usage{
str_format(
  x,
  width = NULL,
  justify = c("left", "right", "centre", "none"),
  ellipsis = "...",
  ...
)
}
\arguments{
\item{x}{a character vector or a string.}

\item{width}{integer. If NULL (default), it is automatically set equal to the max width among all elements of x.}

\item{justify}{one of c("left", "right", "centre", "none").}

\item{ellipsis}{string. It indicates where the string has been truncated.}

\item{...}{other arguments passed to base::format.}
}
\value{
character representations of the elements of x in a common format. THat is, aappropriately padded.
}
\description{
String formatiing/padding
}
