% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{str_pad_num}
\alias{str_pad_num}
\title{Padding of character vector of integers}
\usage{
str_pad_num(x, pad = "0")
}
\arguments{
\item{x}{a numeric vector coercible to integer or it character equivalent.}

\item{pad}{string. Defaults to "0". The padding on the left side.}
}
\value{
a character vector.
}
\description{
Padding of character vector of integers
}
