% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_types.R
\name{symbolic-types}
\alias{symbolic-types}
\alias{symbolic_type0}
\alias{symbolic_type}
\alias{symbolic_types}
\title{Symbolic types}
\usage{
symbolic_type0(x)

symbolic_type(x, abbr = TRUE)

symbolic_types(x, abbr = TRUE)
}
\arguments{
\item{x}{a symbol, a call, or a function.}

\item{abbr}{TRUE or FALSE. If TRUE, the function returns abbreviated symbolic types.}
}
\value{
a string indicating the "symbolic type" of a symbolic object.
}
\description{
Symbolic types
}
