% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_symbolic.R
\name{symbolic_auto_name}
\alias{symbolic_auto_name}
\title{Automatically name symbolic lists}
\usage{
symbolic_auto_name(
  x,
  .name_all = FALSE,
  .names_type = c("full", "sym", "head")
)
}
\arguments{
\item{x}{a list of symbolic objects.}

\item{.names_type}{one of \code{c("full", "sym", "head")}.
\describe{
\item{"full"}{The output is equivalent to \code{deparse(x)}.}
\item{"sym"}{The function returns the string equivalent of a symbol \code{x} is a symbol or a call with no arguments.}
\item{"head"}{The function returns only the string equivalent of a symbol or the call name.}
}}
}
\value{
a named list of symbolics.
}
\description{
Automatically name symbolic lists
}
