% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolic_to_strings.R
\name{symbolics-to-strings}
\alias{symbolics-to-strings}
\alias{symbolic_to_str}
\alias{symbolic_to_chr}
\title{Symbolic to string or character vector}
\usage{
symbolic_to_str(x, .names_type = c("full", "sym", "head"))

symbolic_to_chr(x, .names_type = c("full", "sym", "head"))
}
\arguments{
\item{x}{a symbol, or a call.}

\item{.names_type}{one of \code{c("full", "sym", "head")}.
\describe{
\item{"full"}{The output is equivalent to \code{deparse(x)}.}
\item{"sym"}{The function returns the string equivalent of a symbol \code{x} is a symbol or a call with no arguments.}
\item{"head"}{The function returns only the string equivalent of a symbol or the call name.}
}}
}
\value{
a string for \code{symbolic_to_str} or a character vector for \code{symbolic_to_chr}.
}
\description{
Symbolic to string or character vector
}
